QStringView --> QStringRef substitution example
You might want to recheck uses of QStringView throughout. In general it was straightforward, but a couple tricky things were that the methods 'endswith' and 'startswith' require char instead of string. So when you use a string, I generally called .toString().startsWith(...) etc. Also, they removed the ability to use '==' with QStringView, thus additional calls to toString(). Lastly, it looked like the compiler was ok with adding a 'u' in front of string args but I wasn't sure the implication (e.g., see new line 279 and 287).