Temporarily disable AngleScript for property dialog
Currently, it looks like the ExecuteString function is not passing the appropriate parameters to work properly. This causes all manually entered text to result in a script error and the text to remain unchanged. The only way to change the value is to use the spinner buttons. This can obviously be changed back easily later, but I didn't want to mess with the anglescript engine components.